+++ extensions/zendesk/v1/client/index.ts
@@ -1,2 +1,8 @@
 export { ZendeskAPIClient, makeAPIClient } from './client'
-export type { CreateTicketInput, CreateTicketResponse } from './types'
+export type {
+  CreateTicketInput,
+  CreateTicketResponse,
+  GetTicketResponse,
+  ZendeskTicket,
+  ZendeskUser,
+} from './types'