+++ extensions/zendesk/v1/webhooks/ticketEvent/index.ts
@@ -1,0 +1,1 @@
+export { ticketEvent } from './ticketEvent'